+++ modules/disableAudio.js
@@ -1,10 +1,16 @@
-export default async function disableAudio() {
+import { libraryLog } from "../streamUtils.js";
+
+export default function disableAudio() {
+  libraryLog("Disabling AudioHelper");
   // disables audio since it doesn't work on /stream anyway
   AudioHelper.getAudioContext = function () {
-    if (this._audioContext) return this._audioContext;
     return null;
   };
   AudioHelper.play = function () {
     return;
   };
+  AudioHelper.prototype._onFirstGesture = function () {
+    return;
+  };
+  libraryLog("Finished disabling AudioHelper");
 }