+++ modules/registerHelpers.js
@@ -1,6 +1,8 @@
-import { hasIcon } from "../streamUtils.js";
+import { hasIcon, libraryLog } from "../streamUtils.js";
 
 export default async function registerHelpers() {
+  libraryLog("Registring Handlebars helpers");
+
   Handlebars.registerHelper("getProperty", function (data, property) {
     return getProperty(data, property);
   });
@@ -11,4 +13,6 @@ export default async function registerHelpers() {
   Handlebars.registerHelper("hasIcon", function (dataObject) {
     return hasIcon(dataObject);
   });
+
+  libraryLog("Finished registring Handlebars helpers");
 }